requirements.txt
fastapi 
uvicorn 
python-dotenv
langchain-core 
langchain-openai 
langchain-community 